Course Schedule Improvements (#139)
* Add hyperlinks to lecturers in schedules card

* Allow schedules to be toggled in course search list

---------
