+++ frontend/src/components/CourseSearchList.tsx
@@ -20,6 +20,7 @@ const CoursePage = () => {
 
   const showFCEs = useAppSelector((state) => state.user.showFCEs);
   const showCourseInfos = useAppSelector((state) => state.user.showCourseInfos);
+  const showSchedules = useAppSelector((state) => state.user.showSchedules);
   const loggedIn = useAppSelector((state) => state.user.loggedIn);
 
   const coursesToShow: string[] = useMemo(() => {
@@ -55,6 +56,7 @@ const CoursePage = () => {
             key={course.courseID}
             showFCEs={showFCEs}
             showCourseInfo={showCourseInfos}
+            showSchedules={showSchedules}
           />
         ))}
     </div>
